@@ -16,18 +16,17 @@
         }
     },
     "require": {
-        "php": ">=7.1.0",
-        "psr/container": "~1.0.0"
+        "php": ">=8.0",
+        "psr/container": "~2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~6.2",
-        "squizlabs/php_codesniffer": "~3.0",
-        "phpmd/phpmd": "~2.6",
-        "phan/phan": "~1.1"
+        "phpunit/phpunit": "~9.5",
+        "squizlabs/php_codesniffer": "~3.6",
+        "phan/phan": "~5.3"
     },
     "scripts": {
         "test": "exit 0",
-        "lint": "phpcs --standard=PSR2 --exclude=PSR2.ControlStructures.ControlStructureSpacing src",
-        "analyze": "phpmd src text phpmd.xml; phan --output-mode text --color --allow-polyfill-parser"
+        "lint": "phpcs --standard=PSR12 src",
+        "analyze": "phan --output-mode text --color --no-progress-bar"
     }
 }
